# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Proxmox Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Cluster {
  id: number;
  name: string;
  host: string;
  port: number;
  username: string;
  realm: string;
  status: string;
  vmCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClusterBody {
  name: string;
  host: string;
  port: number;
  username: string;
  password: string;
  realm: string;
}

export interface UpdateClusterBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  realm?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface SyncResult {
  synced: number;
  added: number;
  updated: number;
  message: string;
}

export interface Tenant {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  status: string;
  userCount: number;
  vmCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTenantBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contactEmail?: string | null;
}

export interface UpdateTenantBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface TenantSummary {
  tenantId: number;
  totalVms: number;
  runningVms: number;
  stoppedVms: number;
  pausedVms: number;
  totalUsers: number;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  role: string;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  status: string;
  vmCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateUserBody {
  username: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  role: string;
  /** @nullable */
  tenantId?: number | null;
  password: string;
}

export interface UpdateUserBody {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  password?: string | null;
}

export interface Vm {
  id: number;
  vmId: number;
  name: string;
  node: string;
  type: string;
  status: string;
  /** @nullable */
  cpus?: number | null;
  /** @nullable */
  memoryMb?: number | null;
  /** @nullable */
  diskGb?: number | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  os?: string | null;
  clusterId: number;
  clusterName: string;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateVmBody {
  vmId: number;
  name: string;
  node: string;
  type: string;
  status: string;
  /** @nullable */
  cpus?: number | null;
  /** @nullable */
  memoryMb?: number | null;
  /** @nullable */
  diskGb?: number | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  os?: string | null;
  clusterId: number;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tags?: string | null;
}

export interface UpdateVmBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  cpus?: number | null;
  /** @nullable */
  memoryMb?: number | null;
  /** @nullable */
  diskGb?: number | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tags?: string | null;
}

export interface VmActionBody {
  action: string;
}

export interface VmActionResult {
  success: boolean;
  message: string;
  vmId: number;
  action: string;
}

export interface TenantVmAccess {
  id: number;
  tenantId: number;
  vmId: number;
  tenantName: string;
  vmName: string;
  grantedAt: string;
}

export interface GrantTenantVmAccessBody {
  tenantId: number;
  vmId: number;
}

export interface UserVmAccess {
  id: number;
  userId: number;
  vmId: number;
  userName: string;
  vmName: string;
  grantedAt: string;
}

export interface GrantUserVmAccessBody {
  userId: number;
  vmId: number;
}

export interface DashboardStats {
  totalClusters: number;
  onlineClusters: number;
  totalVms: number;
  runningVms: number;
  stoppedVms: number;
  totalTenants: number;
  activeTenants: number;
  totalUsers: number;
  activeUsers: number;
}

export interface ActivityEvent {
  id: number;
  eventType: string;
  description: string;
  /** @nullable */
  vmId?: number | null;
  /** @nullable */
  vmName?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  createdAt: string;
}

export type ListVmsParams = {
  /**
   * @nullable
   */
  clusterId?: number | null;
  /**
   * @nullable
   */
  tenantId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
};