# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Proxmox Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clusters:
    get:
      operationId: listClusters
      tags: [clusters]
      summary: List all Proxmox clusters
      responses:
        "200":
          description: List of clusters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cluster"
    post:
      operationId: createCluster
      tags: [clusters]
      summary: Register a new Proxmox cluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClusterBody"
      responses:
        "201":
          description: Cluster created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clusters/{id}:
    get:
      operationId: getCluster
      tags: [clusters]
      summary: Get a specific cluster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCluster
      tags: [clusters]
      summary: Update a cluster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClusterBody"
      responses:
        "200":
          description: Cluster updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCluster
      tags: [clusters]
      summary: Remove a cluster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clusters/{id}/sync:
    post:
      operationId: syncCluster
      tags: [clusters]
      summary: Sync VMs from Proxmox cluster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Tenants
...
  /access/tenant-vms:
    get:
      operationId: listTenantVmAccess
      tags: [access]
      summary: List all tenant-VM access grants
      responses:
        "200":
          description: Access list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantVmAccess"
    post:
      operationId: grantTenantVmAccess
      tags: [access]
      summary: Grant a tenant access to a VM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantTenantVmAccessBody"
      responses:
        "201":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantVmAccess"
  /access/tenant-vms/{id}:
    delete:
      operationId: revokeTenantVmAccess
      tags: [access]
      summary: Revoke tenant VM access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
  /access/user-vms:
    get:
      operationId: listUserVmAccess
      tags: [access]
      summary: List all user-VM access grants
      responses:
        "200":
          description: Access list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserVmAccess"
    post:
      operationId: grantUserVmAccess
      tags: [access]
      summary: Grant a user access to a VM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantUserVmAccessBody"
      responses:
        "201":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserVmAccess"
  /access/user-vms/{id}:
    delete:
      operationId: revokeUserVmAccess
      tags: [access]
      summary: Revoke user VM access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Cluster:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        realm:
          type: string
        status:
          type: string
        vmCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - host
        - port
        - username
        - realm
        - status
        - vmCount
        - createdAt
        - updatedAt
    CreateClusterBody:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        realm:
          type: string
      required:
        - name
        - host
        - port
        - username
        - password
        - realm
    UpdateClusterBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        realm:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    SyncResult:
      type: object
      properties:
        synced:
          type: integer
        added:
          type: integer
        updated:
          type: integer
        message:
          type: string
      required:
        - synced
        - added
        - updated
        - message
...
    TenantVmAccess:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        vmId:
          type: integer
        tenantName:
          type: string
        vmName:
          type: string
        grantedAt:
          type: string
      required:
        - id
        - tenantId
        - vmId
        - tenantName
        - vmName
        - grantedAt
    GrantTenantVmAccessBody:
      type: object
      properties:
        tenantId:
          type: integer
        vmId:
          type: integer
      required:
        - tenantId
        - vmId
    UserVmAccess:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        vmId:
          type: integer
        userName:
          type: string
        vmName:
          type: string
        grantedAt:
          type: string
      required:
        - id
        - userId
        - vmId
        - userName
        - vmName
        - grantedAt
    GrantUserVmAccessBody:
      type: object
      properties:
        userId:
          type: integer
        vmId:
          type: integer
      required:
        - userId
        - vmId
...