# lib/api-zod/src/generated/types/tenantVmAccess.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Proxmox Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface TenantVmAccess {
  id: number;
  tenantId: number;
  vmId: number;
  tenantName: string;
  vmName: string;
  grantedAt: string;
}