# lib/api-zod/src/generated/types/userVmAccess.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Proxmox Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UserVmAccess {
  id: number;
  userId: number;
  vmId: number;
  userName: string;
  vmName: string;
  grantedAt: string;
}